+++ lib/time_extension.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/material.dart' show TimeOfDay;
+
+extension TimeOfDayExtension on TimeOfDay {
+  int get totalMinutes => (hour * TimeOfDay.minutesPerHour) + minute;
+
+  bool operator <(TimeOfDay value) => totalMinutes < value.totalMinutes;
+  bool operator <=(TimeOfDay value) => totalMinutes <= value.totalMinutes;
+
+  bool operator >(TimeOfDay value) => totalMinutes > value.totalMinutes;
+  bool operator >=(TimeOfDay value) => totalMinutes >= value.totalMinutes;
+}